chore: fixing the release to use conventionalcommits
By default semantic-release/commit-analyzer does NOT install the
conventionalcommits changelog package necessary to use the
conventionalcommits preset. Since this package wasn't installed, it was
falling back to using the angular one which prevented the previous
commit from releasing the package. This has been fixed.

Additionally, providing a `tsconfig.package.json` for use when
publishing the package. This is due to needing to include tests and
vitest configs in the tsconfig so that TS knows how/where to get the
types.

Also updating the README with ESM syntax in the examples

Also updating the test server handlers to be explicit with the type
since TS was complaining.